agents-intensive-capstone/requirements.txt
# Core dependencies
google-genai>=0.3.0
google-adk>=0.1.0

# MCP Server
mcp>=1.0.0
fastmcp>=0.2.0

# HTTP & API
httpx>=0.27.0
requests>=2.31.0
aiohttp>=3.9.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.11.0

# Environment & Config
python-dotenv>=1.0.0
pydantic>=2.0.0

# Visualization
plotly>=5.18.0
matplotlib>=3.8.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Utilities
python-dateutil>=2.8.0
pytz>=2023.3
